feat: 모달 클릭 시 post
